Split terminal module into directory
